update scripts for running local networks and make sure transfer, register and vote works
